linked all product details, but working on add/update ratings

+++ Cheers_v1/Models/Rating.swift
@@ -30,7 +30,7 @@ struct RatingRow: Identifiable, Codable {
   var id = UUID()
   var rowRating: Double
   var product: String
-  var alc: Int
+  var alc: Double
   var rowPhoto: String
   var style: String
   var dateString: String
